+++ meza/src/meza/pool/pool_ops.cpp
@@ -19,11 +19,11 @@ using qt::u8;
 
 /* ====== ==== */
 void haps_xor_cpu(const meza::pool::matrix_pool<u8> &p,
-		  meza::pool::hap_comp::hap_comp_matrix<u8> &cmp_mat, u32 len,
+		  meza::pool::hap_comp::hap_comp_matrix<u32> &cmp_mat, u32 len,
 		  u32 col_shift, u32 res_shift)
 {
 	const u8 *d_f = p.filter_start_ptr_const();
-	u8 *d_out = cmp_mat.get_xor_data_mut();
+	u32 *d_out = cmp_mat.get_xor_data_mut();
 	u32 p_off = cmp_mat.pool_offset();
 
 	meza::cpu_ops::cpu_haps_xor(d_f, d_out, p_off, col_shift, res_shift,
@@ -31,11 +31,11 @@ void haps_xor_cpu(const meza::pool::matrix_pool<u8> &p,
 }
 
 void haps_sum_cpu(const meza::pool::matrix_pool<u8> &p,
-		  meza::pool::hap_comp::hap_comp_matrix<u8> &cmp_mat, u32 len,
+		  meza::pool::hap_comp::hap_comp_matrix<u32> &cmp_mat, u32 len,
 		  u32 col_shift, u32 res_shift)
 {
 	const u8 *d_f = p.filter_start_ptr_const();
-	u8 *d_out = cmp_mat.get_sum_data_mut();
+	u32 *d_out = cmp_mat.get_sum_data_mut();
 	u32 p_off = cmp_mat.pool_offset();
 
 	meza::cpu_ops::cpu_haps_sum(d_f, d_out, p_off, col_shift, res_shift,
@@ -44,7 +44,7 @@ void haps_sum_cpu(const meza::pool::matrix_pool<u8> &p,
 
 #if MEZA_USE_CUDA
 void haps_xor_cuda(const meza::pool::matrix_pool_cuda<u8> &p,
-		   meza::pool::hap_comp::hap_comp_matrix_cuda<u8> &cmp_mat,
+		   meza::pool::hap_comp::hap_comp_matrix_cuda<u32> &cmp_mat,
 		   u32 len, u32 col_shift, u32 res_shift, cudaStream_t stream)
 {
 	const u8 *d_f = p.filter_ptr();
